Separate AddCommentModal into its own component
