+++ src/components/ObsDetails/ObsDetails.js
@@ -1,9 +1,5 @@
 // @flow
 
-import {
-  BottomSheetModal, BottomSheetModalProvider,
-  BottomSheetTextInput
-} from "@gorhom/bottom-sheet";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import { useQueryClient } from "@tanstack/react-query";
 import { createComment } from "api/comments";
@@ -22,13 +18,13 @@ import _ from "lodash";
 import { ObsEditContext, RealmContext } from "providers/contexts";
 import type { Node } from "react";
 import React, {
-  useContext, useEffect, useRef, useState
+  useContext, useEffect, useState
 } from "react";
 import { useTranslation } from "react-i18next";
 import {
-  Alert, Image, Keyboard,
+  Alert, Image,
   LogBox, Pressable, Text,
-  TextInput as NativeTextInput, TouchableOpacity, View
+  View
 } from "react-native";
 import { ActivityIndicator, Button as IconButton } from "react-native-paper";
 import IconMaterial from "react-native-vector-icons/MaterialIcons";
@@ -42,6 +38,7 @@ import { imageStyles, textStyles, viewStyles } from "styles/obsDetails/obsDetail
 import colors from "styles/tailwindColors";
 
 import ActivityTab from "./ActivityTab";
+import AddCommentModal from "./AddCommentModal";
 import DataTab from "./DataTab";
 import checkCamelAndSnakeCase from "./helpers/checkCamelAndSnakeCase";
 import ObsDetailsHeader from "./ObsDetailsHeader";
@@ -61,16 +58,13 @@ const ObsDetails = ( ): Node => {
   const { t } = useTranslation( );
   const [refetch, setRefetch] = useState( false );
   const [showCommentBox, setShowCommentBox] = useState( false );
-  const [comment, setComment] = useState( "" );
   const { addObservations } = useContext( ObsEditContext );
   const { params } = useRoute( );
   const { uuid } = params;
   const [tab, setTab] = useState( 0 );
   const navigation = useNavigation( );
   const [ids, setIds] = useState( [] );
-  const bottomSheetModalRef = useRef( null );
   const [addingComment, setAddingComment] = useState( false );
-  const [snapPoint, setSnapPoint] = useState( 100 );
   const realm = useRealm( );
   const localObservation = realm?.objectForPrimaryKey( "Observation", uuid );
 
@@ -144,50 +138,6 @@ const ObsDetails = ( ): Node => {
     }
   }, [observation, ids] );
 
-  // Clear the comment in a timeout so it doesn't trigger a re-render of the
-  // text input *after* the bottom sheet modal gets dismissed, b/c that seems
-  // to re-render the bottom sheet in a presented state, making it hard to
-  // actually dismiss
-  const clearComment = ( ) => setTimeout( ( ) => setComment( "" ), 100 );
-
-  const onBackdropPress = () => {
-    Alert.alert(
-      t( "Discard-Comment" ),
-      t( "Are-you-sure-discard-comment" ),
-      [{
-        text: t( "Yes" ),
-        onPress: () => {
-          setShowCommentBox( false );
-          // setComment( "" );
-          Keyboard.dismiss();
-          clearComment( );
-        }
-      }, { text: t( "No" ) }],
-      {
-        cancelable: false
-      }
-    );
-  };
-
-  // Make bottom sheet modal visibility reactive instead of imperative
-  useEffect( ( ) => {
-    if ( showCommentBox ) {
-      bottomSheetModalRef.current?.present( );
-    } else {
-      bottomSheetModalRef.current?.dismiss( );
-    }
-  }, [showCommentBox, bottomSheetModalRef] );
-
-  const renderHandle = () => (
-    <View style={viewStyles.handleContainer} />
-  );
-
-  const renderBackdrop = () => (
-    <TouchableOpacity activeOpacity={1} style={viewStyles.background} onPress={onBackdropPress}>
-      <View />
-    </TouchableOpacity>
-  );
-
   const showActivityTab = ( ) => setTab( 0 );
   const showDataTab = ( ) => setTab( 1 );
 
@@ -246,19 +196,7 @@ const ObsDetails = ( ): Node => {
     addObservations( [observation] );
     navigation.push( "AddID", { onIDAdded, goBackOnSave: true } );
   };
-  const openCommentBox = ( ) => {
-    setShowCommentBox( true );
-  };
-  const submitComment = async ( ) => {
-    setAddingComment( true );
-    clearComment( );
-    setShowCommentBox( false );
-    Keyboard.dismiss();
-    if ( comment.length > 0 ) {
-      createCommentMutation.mutate( comment );
-    }
-    setAddingComment( false );
-  };
+  const openCommentBox = ( ) => setShowCommentBox( true );
 
   const showTaxon = ( ) => {
     if ( !taxon ) { return <Text>{t( "Unknown-organism" )}</Text>; }
@@ -281,30 +219,6 @@ const ObsDetails = ( ): Node => {
     );
   };
 
-  const renderBottomSheetTextView = () => (
-    <BottomSheetTextInput
-      keyboardType="default"
-      style={[viewStyles.commentInput, viewStyles.commentInputText, textStyles.commentTextInput]}
-      value={comment}
-      selectionColor={colors.black}
-      activeUnderlineColor={colors.transparent}
-      placeholder={t( "Add-a-comment" )}
-      autoFocus
-      multiline
-      onChangeText={setComment}
-      render={innerProps => (
-        <NativeTextInput
-              /* eslint-disable react/jsx-props-no-spreading */
-          {...innerProps}
-          style={[
-            innerProps.style,
-            viewStyles.commentInputText, textStyles.commentTextInput
-          ]}
-        />
-      )}
-    />
-  );
-
   const faveOrUnfave = async ( ) => {
     // TODO: fix fave/unfave functionality with useMutation
     if ( currentUserFaved ) {
@@ -323,7 +237,7 @@ const ObsDetails = ( ): Node => {
     : formatObsListTime( observation._created_at ) );
 
   return (
-    <BottomSheetModalProvider>
+    <>
       <ScrollWithFooter testID={`ObsDetails.${uuid}`}>
         <ObsDetailsHeader observation={observation} />
         <View style={viewStyles.userProfileRow}>
@@ -410,9 +324,9 @@ const ObsDetails = ( ): Node => {
           )
           : <DataTab observation={observation} />}
         {addingComment && (
-        <View style={[viewStyles.row, viewStyles.centerRow]}>
-          <ActivityIndicator size="large" />
-        </View>
+          <View style={[viewStyles.row, viewStyles.centerRow]}>
+            <ActivityIndicator size="large" />
+          </View>
         )}
         <View style={viewStyles.row}>
           <View style={viewStyles.buttons}>
@@ -434,36 +348,13 @@ const ObsDetails = ( ): Node => {
           </View>
         </View>
       </ScrollWithFooter>
-      <BottomSheetModal
-        ref={bottomSheetModalRef}
-        index={0}
-        enableOverDrag={false}
-        enablePanDownToClose
-        snapPoints={[snapPoint]}
-        backdropComponent={renderBackdrop}
-        handleComponent={renderHandle}
-        style={viewStyles.bottomModal}
-      >
-        <View
-          style={viewStyles.commentInputContainer}
-          onLayout={( {
-            nativeEvent: {
-              layout: { height }
-            }
-          } ) => {
-            setSnapPoint( height + 20 );
-          }}
-        >
-          {renderBottomSheetTextView()}
-          <TouchableOpacity
-            style={viewStyles.sendComment}
-            onPress={() => submitComment( )}
-          >
-            <IconMaterial name="send" size={35} color={colors.inatGreen} />
-          </TouchableOpacity>
-        </View>
-      </BottomSheetModal>
-    </BottomSheetModalProvider>
+      <AddCommentModal
+        createCommentMutation={createCommentMutation}
+        showCommentBox={showCommentBox}
+        setShowCommentBox={setShowCommentBox}
+        setAddingComment={setAddingComment}
+      />
+    </>
   );
 };
